+++ src/main/java/ru/skypro/homework/service/CommentService.java
@@ -34,7 +34,7 @@ public interface CommentService {
      * @param adId      идентификатор объявления
      * @param commentId идентификатор комментария
      */
-    void rmComment(int adId, int commentId);
+    void rmComment(Long adId, Long commentId);
 
     /**
      * Обновляет комментарий
